+++ core/sfc-core/src/main/kotlin/com/amazonaws/sfc/util/WorkerQueue.kt
@@ -4,7 +4,6 @@ import com.amazonaws.sfc.log.Logger
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
 import kotlin.coroutines.CoroutineContext
-import kotlin.time.Duration
 
 class WorkerQueue<T, R>(
     private val workers: Int,