+++ targets/opcua-target/src/main/kotlin/com/amazonaws/sfc/opcuatarget/OpcuaServerDataTypes.kt
@@ -5,11 +5,12 @@
 package com.amazonaws.sfc.opcuatarget
 
 
-import com.amazonaws.sfc.data.JsonHelper
+import com.amazonaws.sfc.config.ChannelConfiguration.Companion.CONFIG_TRANSFORMATION
+import com.amazonaws.sfc.log.Logger
+import com.amazonaws.sfc.opcuatarget.OpcuaDataTypes.Companion.asType
 import org.eclipse.milo.opcua.stack.core.Identifiers
 import org.eclipse.milo.opcua.stack.core.types.builtin.*
 import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UInteger
-import kotlin.math.truncate
 
 enum class OpcuaServerDataTypes {
 
@@ -78,7 +79,7 @@ enum class OpcuaServerDataTypes {
         override val identifier: NodeId
             get() = Identifiers.QualifiedName
     },
-    REAL{
+    REAL {
         override val identifier: NodeId
             get() = Identifiers.Float
     },
@@ -150,96 +151,50 @@ enum class OpcuaServerDataTypes {
     abstract val identifier: NodeId
 
     companion object {
+
+        private val className = this::class.java.name.toString()
+
         fun fromString(value: String): OpcuaServerDataTypes {
             val s = value.uppercase().trim().replace("_", "")
             return entries.find { it.name == s } ?: UNDEFINED
         }
 
 
-        private fun convert(value: Any?, dataTypeIdentifier: NodeId?, dimensions: List<Int>?): Any? {
+        private fun convert(value: Any?, dataTypeIdentifier: NodeId?, dimensions: List<Int>?, logger: Logger): Any? {
             var v = value
             try {
-                if (v is List<*> ) {
+                if (v is List<*> && dimensions != null) {
 
                     v = when (dataTypeIdentifier) {
-
-                        Identifiers.Boolean ->
-                            deepCast<Boolean>(dimensions, v) { it as Boolean }
-
-                        Identifiers.SByte ->
-                            deepCast<Byte>(dimensions, v) { it as Byte }
-
-                        Identifiers.ByteString ->
-                            deepCast<ByteString>(dimensions, v) {
-                                ByteString.of(it.toString().encodeToByteArray())
-                            }
-
-                        Identifiers.String ->
-                            deepCast<String>(dimensions, v) { it as String }
-
-                        Identifiers.Structure ->
-                            deepCast<String>(dimensions, v) { JsonHelper.gsonExtended().toJson(it) }
-
-                        Identifiers.DateTime ->
-                            deepCast<DateTime>(dimensions, v) { it as DateTime }
-
-                        Identifiers.Double ->
-                            deepCast<Double>(dimensions, v) { it as Double }
-
-                        Identifiers.ExpandedNodeId ->
-                            deepCast<ExpandedNodeId>(dimensions, v) { it as ExpandedNodeId }
-
-                        Identifiers.Float ->
-                            deepCast<Float>(dimensions, v) { it as Float }
-
-                        Identifiers.Int16 ->
-                            deepCast<Short>(dimensions, v) { it as Short }
-
-                        Identifiers.Int32 ->
-                            deepCast<Int>(dimensions, v) { it as Int }
-
-                        Identifiers.Int64 ->
-                            deepCast<Long>(dimensions, v) { it as Long }
-
-                        Identifiers.Byte ->
-                            deepCast<org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UByte>(
-                                dimensions,
-                                v) { org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UByte.valueOf(it as Byte) }
-
-                        Identifiers.UInt16 ->
-                            deepCast<org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UShort>(
-                                dimensions,
-                                v) { org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UShort.valueOf(it as Short) }
-
-                        Identifiers.UInt32 ->
-                            deepCast<UInteger>(dimensions, v) { UInteger.valueOf(it as Int) }
-
-                        Identifiers.UInt64 ->
-                            deepCast<org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.ULong>(
-                                dimensions,
-                                v) { org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.ULong.valueOf(it as Long) }
-
-                        Identifiers.NodeId ->
-                            deepCast<NodeId>(dimensions, v) { it as NodeId }
-
-                        Identifiers.XmlElement ->
-                            deepCast<XmlElement>(dimensions, v) { XmlElement(it.toString()) }
-
-                        else -> {
-                            v
-                        }
-                    }
-                } else{
-                    return when(v){
-                        is UByte -> v.toShort()
-                        is UShort -> v.toInt()
-                        is UInt -> v.toLong()
-                        is ULong -> truncate(v.toDouble())
+                        Identifiers.Boolean -> deepCast<Boolean>(dimensions, v) { asType(it, Identifiers.Boolean) as Boolean }
+                        Identifiers.SByte -> deepCast<Byte>(dimensions, v) { asType(it, Identifiers.SByte) as Byte }
+                        Identifiers.ByteString -> deepCast<ByteString>(dimensions, v) { asType(it, Identifiers.ByteString) as ByteString }
+                        Identifiers.String -> deepCast<String>(dimensions, v) { asType(it, Identifiers.String) as String }
+                        Identifiers.Structure -> deepCast<String>(dimensions, v) { asType(it, Identifiers.Structure) as String }
+                        Identifiers.DateTime -> deepCast<DateTime>(dimensions, v) { asType(it, Identifiers.DateTime) as DateTime }
+                        Identifiers.Double -> deepCast<Double>(dimensions, v) { asType(it, Identifiers.Double) as Double }
+                        Identifiers.ExpandedNodeId -> deepCast<ExpandedNodeId>(dimensions, v) { asType(it, Identifiers.ExpandedNodeId) as ExpandedNodeId }
+                        Identifiers.Float -> deepCast<Float>(dimensions, v) { asType(it, Identifiers.Float) as Float }
+                        Identifiers.Int16 -> deepCast<Short>(dimensions, v) { asType(it, Identifiers.Int16) as Short }
+                        Identifiers.Int32 -> deepCast<Int>(dimensions, v) { asType(it, Identifiers.Int32) as Int }
+                        Identifiers.Int64 -> deepCast<Long>(dimensions, v) { asType(it, Identifiers.Int64) as Long }
+                        Identifiers.Byte -> deepCast<UByte>(dimensions, v) { asType(it, Identifiers.Byte) as UByte }
+                        Identifiers.UInt16 -> deepCast<UShort>(dimensions, v) { asType(it, Identifiers.Int32) as UShort }
+                        Identifiers.UInt32 -> deepCast<UInteger>(dimensions, v) { UInteger.valueOf(asType(it, Identifiers.Int64) as Long) }
+                        Identifiers.UInt64 -> deepCast<ULong>(dimensions, v) { asType(it, Identifiers.Int64) as ULong }
+                        Identifiers.NodeId -> deepCast<NodeId>(dimensions, v) { asType(it, Identifiers.NodeId) as NodeId }
+                        Identifiers.XmlElement -> deepCast<XmlElement>(dimensions, v) { asType(it, Identifiers.XmlElement) as XmlElement }
                         else -> v
                     }
-
+                } else {
+                    v = if (v != null) {
+                        asType(v, dataTypeIdentifier!!)
+                    } else v
                 }
             } catch (_: Exception) {
+                logger.getCtxErrorLog(
+                    className,
+                    "convert")("Error converting value $value:${value!!::class.simpleName} to OPCUA data type $dataTypeIdentifier, set the $CONFIG_TRANSFORMATION property of the node configuration to convert to the required type")
                 return null
             }
             return v
@@ -281,7 +236,7 @@ enum class OpcuaServerDataTypes {
             }
         }
 
-        fun Any?.toVariant(dataTypeIdentifier: NodeId?, dimensions: List<Int>?): Variant = Variant(convert(this, dataTypeIdentifier, dimensions))
+        fun Any?.toVariant(dataTypeIdentifier: NodeId?, dimensions: List<Int>?, logger : Logger): Variant = Variant(convert(this, dataTypeIdentifier, dimensions, logger))
     }
 
 }